requirements.txt
amocrm-api==2.3.1
APScheduler==3.6.3
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
flow==0.0.1
fpdf==1.7.2
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
oauth2client==4.1.3
oauthlib==3.1.1
packaging==20.9
pendulum==2.1.2
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==2.1.0
pyparsing==2.4.7
pyTelegramBotAPI==3.7.9
python-dateutil==2.8.1
python-slugify==5.0.2
python-telegram-bot==13.6
pytz==2021.1
pytzdata==2020.1
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
slugify==0.0.1
text-unidecode==1.3
tornado==6.1
tzlocal==2.1
uritemplate==3.0.1
urllib3==1.26.5
